intake-tool/requirements.txt
accelerate
aext-assistant
aext-assistant-server
aext-core
aext-core-server
aext-panels
aext-panels-server
aext-share-notebook
aext-share-notebook-server
aext-shared
aiobotocore
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
alabaster
altair
anaconda-anon-usage
anaconda-catalogs
anaconda-client
anaconda-cloud-auth
anaconda-navigator
anaconda-project
annotated-types
any
anyio
appdirs
APScheduler
archspec
argon2-cffi
argon2-cffi-bindings
arrow
astroid
astropy
astropy-iers-data
asttokens
async-lru
atomicwrites
attrs
Automat
autopep8
Babel
bcrypt
beautifulsoup4
binaryornot
black
bleach
blinker
blis
bokeh
boltons
botocore
Bottleneck
Brotli
cachetools
catalogue
Cerberus
certifi
cffi
chardet
charset-normalizer
click
cloudpathlib
cloudpickle
colorama
colorcet
comm
conda
conda-build
conda-content-trust
conda-libmamba-solver
conda-pack
conda-package-handling
conda-repo-cli
conda-token
conda_index
conda_package_streaming
confection
constantly
contourpy
cookiecutter
cryptography
cssselect
cycler
cymem
cytoolz
dask
dask-expr
datashader
debugpy
decorator
defusedxml
Deprecated
diff-match-patch
dill
distributed
distro
docstring-to-markdown
docstring_parser
docutils
en-core-web-sm
et-xmlfile
executing
fastapi
fastjsonschema
filelock
flake8
Flask
Flask-APScheduler
Flask-Cors
Flask-Limiter
fonttools
frozendict
frozenlist
fsspec
gensim
gitdb
GitPython
google-api-core
google-auth
google-cloud-aiplatform
google-cloud-appengine-logging
google-cloud-audit-log
google-cloud-bigquery
google-cloud-core
google-cloud-logging
google-cloud-resource-manager
google-cloud-secret-manager
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpc-google-iam-v1
grpcio
grpcio-status
h11
h5py
HeapDict
holoviews
httpcore
httpx
huggingface-hub
hvplot
hyperlink
idna
imagecodecs
imageio
imagesize
imbalanced-learn
importlib-metadata
importlib_resources
incremental
inflection
iniconfig
intake
intervaltree
ipykernel
ipython
ipython-genutils
ipywidgets
isort
itemadapter
itemloaders
itsdangerous
jaraco.classes
jedi
jellyfish
Jinja2
jiter
jmespath
joblib
json5
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
keyring
kiwisolver
langcodes
language_data
lazy-object-proxy
lazy_loader
lckr_jupyterlab_variableinspector
libarchive-c
libmambapy
limits
linkify-it-py
llvmlite
lmdb
locket
lxml
lz4
marisa-trie
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mccabe
mdit-py-plugins
mdurl
menuinst
mistune
mkl-service
mkl_fft
mkl_random
more-itertools
mpmath
msgpack
multidict
multipledispatch
murmurhash
mypy
mypy-extensions
navigator-updater
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
notebook_shim
numba
numexpr
numpy
numpydoc
openai
openpyxl
opentelemetry-api
ordered-set
overrides
packaging
pandas
pandocfilters
panel
param
paramiko
parsel
parso
partd
pathspec
patsy
pdf2image
pexpect
pickleshare
pillow
pipenv
pkce
pkginfo
platformdirs
plotly
pluggy
ply
preshed
prometheus-client
prompt-toolkit
Protego
proto-plus
protobuf
psutil
ptyprocess
pure-eval
py-cpuinfo
pyarrow
pyasn1
pyasn1-modules
pycodestyle
pycosat
pycparser
pyct
pycurl
pydantic
pydantic_core
pydeck
PyDispatcher
pydocstyle
pyerfa
pyflakes
Pygments
PyJWT
pylint
pylint-venv
pyls-spyder
PyNaCl
pyodbc
pyOpenSSL
pyparsing
PyQt5
PyQt5-sip
PyQtWebEngine
PySocks
pytesseract
pytest
python-dateutil
python-dotenv
python-json-logger
python-lsp-black
python-lsp-jsonrpc
python-lsp-server
python-slugify
pytoolconfig
pytz
pyviz_comms
PyWavelets
pywin32
pywin32-ctypes
pywinpty
PyYAML
pyzmq
QDarkStyle
qstylizer
QtAwesome
qtconsole
QtPy
queuelib
ray
redis
referencing
regex
reportlab
requests
requests-file
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rope
rpds-py
rsa
Rtree
ruamel-yaml-conda
ruamel.yaml
ruamel.yaml.clib
s3fs
safetensors
scikit-image
scikit-learn
scipy
Scrapy
seaborn
semver
Send2Trash
sentencepiece
service-identity
setuptools
shapely
shellingham
sip
six
smart-open
smmap
sniffio
snowballstemmer
sortedcontainers
soupsieve
spacy
spacy-legacy
spacy-loggers
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
SQLAlchemy
srsly
stack-data
starlette
statsmodels
streamlit
sympy
tables
tabulate
tblib
tenacity
terminado
text-unidecode
textdistance
thinc
threadpoolctl
three-merge
tifffile
tinycss2
tldextract
tokenizers
toml
tomli
tomlkit
toolz
torch
torchaudio
torchvision
tornado
tqdm
traitlets
transformers
truststore
Twisted
twisted-iocpsupport
typer
typing_extensions
tzdata
tzlocal
uc-micro-py
ujson
unicodedata2
Unidecode
urllib3
uvicorn
w3lib
wasabi
watchdog
wcwidth
weasel
webencodings
websocket-client
Werkzeug
whatthepatch
wheel
widgetsnbextension
win-inet-pton
wrapt
xarray
xlwings
xyzservices
yapf
yarl
zict
zipp
zope.interface
zstandard
